+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -2655,7 +2655,7 @@ export const HTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6AKQCPXKBE0QCRMSN5AN41E"),
+  _id: z.string().default("01K6ASAWK46065DPKG7P4JAC7E"),
   key: z.string(),
   display_name: z.string(),
   description: z.string(),
@@ -2688,7 +2688,7 @@ export const HTTPTool$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   HTTPTool
 > = z.object({
-  id: z.string().default("01K6AKQCPXKBE0QCRMSN5AN41E"),
+  id: z.string().default("01K6ASAWK46065DPKG7P4JAC7E"),
   key: z.string(),
   displayName: z.string(),
   description: z.string(),