+++ packages/orq-rc/RELEASES.md
@@ -3919,3 +3919,13 @@ Based on:
 - [typescript v3.13.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.44 - packages/orq-rc
+
+## 2025-09-29 13:08:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.4 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.45 - packages/orq-rc